so made the board for the order and ordermanager
board- dont have textures and i dont know how to keep them on it, please do take a look and leave me a message if you know how to add a color or texture

ordermanager- coded abit but will be completed as we finish the order pick up mechanism as shown in the video.

please take a look and let me know add the board to the game scene and see if it fits and resize according to the wall size.